requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
pandas==2.0.3
pillow==10.1.0
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
ratelimit==2.2.1 